Fix memory leak due to adding interfaces several times to the cache.
getifaddrs() may return an interface several times (if it has several
IP). We handle this case gracefully by not freeing the allocated
resources when they are not needed and trying the next interface as
soon as the case is detected.

A proper fix will be to handle each interface once.